kafka_consumer/requirements.txt
azure-storage-blob==12.24.0
# configparser==7.1.0
pandas==2.2.3
confluent-kafka==2.6.1
boto3==1.35.73